[no ci] moved package module into package.rs
